+++ test/vhost/migration/migration-tc3a.sh
@@ -100,7 +100,7 @@ function host1_start_nvmf()
 
 	notice "Starting nvmf_tgt instance on local server"
 	mkdir -p $nvmf_dir
-	rm -rf $nvmf_dir/*
+	rm -rf "${nvmf_dir:?}/"*
 
 	trap 'host1_cleanup_nvmf SIGKILL; error_exit "${FUNCNAME}" "${LINENO}"' INT ERR EXIT
 	$rootdir/app/nvmf_tgt/nvmf_tgt -s 512 -m 0xF -r $nvmf_dir/nvmf_rpc.sock --wait-for-rpc &
